Based on feedback from thom store individual github org contributions
This allows us to run a single bot against multiple orgs. Example:

User contributes to chef and gets the welcome message, and then contributes to chef-cookbooks and gets a different welcome message as the process there is different. At the moment the message is still the same generic message, but this gets the data store like we need it. We’re just storing a big JSON blog now since there isn’t a hash type for awe-record. It works. It’s less than ideal since we have to parse it out every time to try to add values.

Signed-off-by: Tim Smith <tsmith@chef.io>